Improve code

+++ three/core/renderer/cpu/ray_tracing/renderer.h
@@ -10,6 +10,6 @@ public:
     void render(std::shared_ptr<Scene> scene,
         std::shared_ptr<Camera> camera,
         std::shared_ptr<RayTracingOptions> options,
-        pybind11::array_t<int, pybind11::array::c_style> buffer);
+        pybind11::array_t<float, pybind11::array::c_style> buffer);
 };
 }
